StyleCop refactorings.

+++ LibiadaCore/Core/Characteristics/Calculators/ArithmeticMean.cs
@@ -32,8 +32,8 @@ namespace LibiadaCore.Core.Characteristics.Calculators
         public double Calculate(CongenericChain chain, Link link)
         {
             double sum = adder.Calculate(chain, link);
-            double intervalsCount = counter.Calculate(chain, link);
-            return intervalsCount == 0 ? 0: sum / intervalsCount;
+            int intervalsCount = (int)counter.Calculate(chain, link);
+            return intervalsCount == 0 ? 0 : sum / intervalsCount;
         }
 
         /// <summary>
