+++ LibiadaCore/Core/Characteristics/Calculators/AverageRemotenessSkewness.cs
@@ -31,7 +31,7 @@
         {
             double result = 0;
             double g = averageRemoteness.Calculate(chain, link);
-            double n = intervalsCount.Calculate(chain, link);
+            int n = (int)intervalsCount.Calculate(chain, link);
 
             for (int i = 0; i < chain.Alphabet.Cardinality; i++)
             {
